Make package arch any, adjust plugin-registry according to build arch
Audacious is a bit shit, shipping the settings this way is a bit shit,
just trying to make the best of a crappy situation.

Audacious needs a properly configured plugin registry in order to be
able to load the settings. The plugin registry however is a runtime
built cache and uses full paths to .so's making the registry invalid
on anything but x86_64. So, in order to ship the settings properly we're
going to sed out markers with the proper arch for every arch.